requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
Authlib==1.3.1
Gunicorn==22.0.0
